Issue #4198: Introduce new FilesetsTrait
...and move all protected filesets helper methods to it
